unification of results in between lazarus and delphi, adjustment for …

+++ generator/Common/generate.console.pas
@@ -44,7 +44,7 @@ resourcestring
 var
   inputFilename: String = '';
   outputFilename: String = '';
-  lineCount: Int64 = 0;
+  lineCount: Integer = 0;
 
 procedure WriteHelp;
